+++ modules/graphics/src/main/java/com/sun/javafx/tk/quantum/GlassTouchEventListener.java
@@ -30,8 +30,8 @@ interface GlassTouchEventListener {
     public void notifyBeginTouchEvent(long time, int modifiers, boolean isDirect,
             int touchEventCount);
 
-    public void notifyNextTouchEvent(long time, int type, long touchId, 
-                                     double x, double y, double xAbs, double yAbs);
+    public void notifyNextTouchEvent(long time, int type, long touchId,
+                                     int x, int y, int xAbs, int yAbs);
 
     public void notifyEndTouchEvent(long time);
 }
